dv360-reports-automation/dv360_reports_download/requirements.txt
google-cloud==0.34.0
pandas==1.3.5
requests==2.28.0
google-cloud-bigquery==3.2.0
google-cloud-storage==2.4.0
oauth2client
pandas-gbq
google-api-python-client